Split Dozent Controller
